💄 feat(sharedUI): adaptive chapter panel in commonMain reusing PlayerChapterRow
