--- sharedUI/src/androidMain/kotlin/com/calypsan/listenup/client/features/nowplaying/ChapterPickerSheet.kt
@@ -1,174 +1,0 @@
-package com.calypsan.listenup.client.features.nowplaying
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.navigationBarsPadding
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import com.calypsan.listenup.client.domain.model.Chapter
-import kotlin.time.Duration.Companion.milliseconds
-
-/**
- * Bottom sheet for selecting a chapter to jump to.
- *
- * Shows all chapters with their durations, highlighting the current chapter
- * with a play icon indicator.
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ChapterPickerSheet(
-    chapters: List<Chapter>,
-    currentChapterIndex: Int,
-    onChapterSelected: (Int) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true),
-        shape = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp),
-        dragHandle = {
-            // Standard drag handle with proper spacing
-            Surface(
-                modifier =
-                    Modifier
-                        .padding(vertical = 12.dp)
-                        .width(32.dp)
-                        .height(4.dp),
-                shape = RoundedCornerShape(2.dp),
-                color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.4f),
-            ) {}
-        },
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .navigationBarsPadding(),
-        ) {
-            // Header
-            Text(
-                text = "Chapters",
-                style = MaterialTheme.typography.titleLarge,
-                modifier = Modifier.padding(horizontal = 24.dp, vertical = 8.dp),
-            )
-
-            HorizontalDivider()
-
-            // Chapter list
-            LazyColumn(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .weight(1f, fill = false),
-            ) {
-                itemsIndexed(chapters) { index, chapter ->
-                    ChapterRow(
-                        chapter = chapter,
-                        index = index,
-                        isCurrentChapter = index == currentChapterIndex,
-                        onClick = { onChapterSelected(index) },
-                    )
-                }
-            }
-
-            Spacer(Modifier.height(16.dp))
-        }
-    }
-}
-
-@Composable
-private fun ChapterRow(
-    chapter: Chapter,
-    index: Int,
-    isCurrentChapter: Boolean,
-    onClick: () -> Unit,
-) {
-    Surface(
-        onClick = onClick,
-        color =
-            if (isCurrentChapter) {
-                MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
-            } else {
-                Color.Transparent
-            },
-    ) {
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 24.dp, vertical = 14.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            // Chapter number
-            Text(
-                text = "${index + 1}",
-                style = MaterialTheme.typography.labelLarge,
-                color =
-                    if (isCurrentChapter) {
-                        MaterialTheme.colorScheme.primary
-                    } else {
-                        MaterialTheme.colorScheme.onSurfaceVariant
-                    },
-                modifier = Modifier.width(32.dp),
-            )
-
-            // Chapter title
-            Text(
-                text = chapter.title,
-                style = MaterialTheme.typography.bodyLarge,
-                color =
-                    if (isCurrentChapter) {
-                        MaterialTheme.colorScheme.primary
-                    } else {
-                        MaterialTheme.colorScheme.onSurface
-                    },
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis,
-                modifier = Modifier.weight(1f),
-            )
-
-            Spacer(Modifier.width(12.dp))
-
-            // Duration
-            Text(
-                text = chapter.duration.milliseconds.formatPlaybackTime(),
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-
-            // Playing indicator
-            if (isCurrentChapter) {
-                Spacer(Modifier.width(8.dp))
-                Icon(
-                    Icons.Default.PlayArrow,
-                    contentDescription = "Now playing",
-                    tint = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.size(20.dp),
-                )
-            }
-        }
-    }
-}
